requirements.txt
cryptography>=42.0.0
pycryptodome>=3.20.0
tkinter-tooltip>=2.1.0
